Adds new features: - Role - Permission - User Access Control - Deactivate and remove inactive user - Audit

+++ src/controllers/editionController.ts
@@ -1,6 +1,7 @@
 import logger from '@config/logger'
 import { EditionService } from '@services/EditionService'
 import express from 'express'
+import { can } from 'src/middleware/authorization.middleware'
 import httpMiddleware from 'src/middleware/http.middleware'
 
 const editionRoutes = express.Router()
@@ -8,7 +9,7 @@ editionRoutes.use(httpMiddleware)
 
 const editionService = new EditionService()
 
-editionRoutes.get('/model/:model', async (req: any, res: any) => {
+editionRoutes.get('/model/:model', can([]), async (req: any, res: any) => {
   const modelName = req.params.model
   logger.info('GET /' + req.params?.model)
   try {
@@ -22,14 +23,14 @@ editionRoutes.get('/model/:model', async (req: any, res: any) => {
   }
 })
 
-editionRoutes.get('/model/:model/:param', async (req: any, res: any) => {
+editionRoutes.get('/model/:model/:param', can([]), async (req: any, res: any) => {
   const modelName = req.params.model
   const params = req.query
   logger.info(`GET /model/${modelName}/${JSON.stringify(params)}`)
   res.status(200).json(await editionService.find({ model: modelName, data: params }))
 })
 
-editionRoutes.get('/model-attributes/:model', async (req: any, res: any) => {
+editionRoutes.get('/model-attributes/:model', can([]), async (req: any, res: any) => {
   const modelName = req.params.model
   logger.info(`GET /model-attributes/${modelName}`)
   res.status(200).json(await editionService.getAttributes(modelName))
@@ -40,7 +41,7 @@ editionRoutes.get('/menu-options', async (req: any, res: any) => {
   res.status(200).json(await editionService.getMenuOptions())
 })
 
-editionRoutes.post('/model/:model/search', async (req: any, res: any) => {
+editionRoutes.post('/model/:model/search', can([]), async (req: any, res: any) => {
   const modelName = req.params.model
   const searchOptions = req.query
   logger.info(`POST /model/${modelName}/search`)
@@ -51,32 +52,32 @@ editionRoutes.post('/model/:model/search', async (req: any, res: any) => {
   }
 })
 
-editionRoutes.post('/model/:model', async (req: any, res: any) => {
+editionRoutes.post('/model/:model', can([]), async (req: any, res: any) => {
   const modelName = req.params.model
   logger.info(`POST /model/${modelName}`)
   try {
-    res.status(200).json(await editionService.create({ model: modelName, data: req.body }))
+    res.status(200).json(await editionService.create({ model: modelName, data: req.body, userId: req.userId }))
   } catch (error) {
     return res.status(400).json(error)
   }
 })
 
-editionRoutes.put('/model/:model', async (req, res) => {
+editionRoutes.put('/model/:model', can([]), async (req: any, res: any) => {
   const modelName = req.params.model
   logger.info(`PUT /model/${modelName}`)
   try {
-    res.status(200).json(await editionService.update({ model: modelName, data: req.body }))
+    res.status(200).json(await editionService.update({ model: modelName, data: req.body, userId: req.userId }))
   } catch (error) {
     return res.status(400).json(error)
   }
 })
 
-editionRoutes.delete('/model/:model/:param', async (req, res) => {
+editionRoutes.delete('/model/:model/:param', can([]), async (req: any, res: any) => {
   const modelName = req.params.model
   const param = req.params.param
   logger.info(`DELETE /model/${modelName}/${param}`)
   try {
-    res.status(200).json(await editionService.delete({ model: modelName, data: param }))
+    res.status(200).json(await editionService.delete({ model: modelName, data: param, userId: req.userId }))
   } catch (error) {
     return res.status(400).json(error)
   }
